requirements.txt
langgraph>=0.2.0
openai>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
lxml>=4.9.0
newspaper3k>=0.2.8
nltk>=3.8.0
